requirements.txt
albucore==0.0.21
albumentations==1.4.22
annotated-types==0.7.0
contourpy==1.3.1
cycler==0.12.1
eval_type_backport==0.2.0
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.10.0
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.4.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.2
pillow==11.0.0
pydantic==2.10.3
pydantic_core==2.27.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
scipy==1.14.1
simsimd==6.2.1
six==1.17.0
stringzilla==3.11.1
sympy==1.13.1
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
typing_extensions==4.12.2
